wasmer_sdk::app::deploy_remote_build

Function deploy_app_remote

pub async fn deploy_app_remote<F>(
    client: &WasmerClient,
    opts: DeployRemoteOpts,
    base_dir: &Path,
    on_progress: F,
) -> Result<DeployAppVersion, DeployRemoteError>
where
    F: FnMut(DeployRemoteEvent) + Send,

Deploy an application using the remote autobuild zip upload flow.

It will build a ZIP archive of the specified base_dir, upload it to Wasmer, and request an autobuild deployment.

The Wasmer backend will handle building and deploying the application.